@@ -2,6 +2,7 @@ import { SocialAuthService } from "@abacritt/angularx-social-login";
 import { Component, Input } from "@angular/core";
 import { Router } from "@angular/router";
 import { PreferencesService } from "../../services/preferences.service";
+import { LoginService } from '../../../modules/login/services/login.service';
 
 @Component({
     selector: 'enarm-header',
@@ -17,14 +18,16 @@ export class HeaderComponent {
     constructor(
         public router: Router,
         public preferencesService: PreferencesService,
-        public socialAuthService: SocialAuthService
+        public socialAuthService: SocialAuthService,
+        public loginService: LoginService
     ) {
 
     }
 
     async logout() {
-        await this.preferencesService.clearAllItems();
         try {
+            await this.loginService.logout()
+            await this.preferencesService.clearAllItems();
             await this.socialAuthService.signOut()
             this.router.navigateByUrl('login');
 
