@@ -16,7 +16,7 @@ export class ProgressBarExamComponent implements OnInit {
     }
 
     async ngOnInit(): Promise<void> {
-        
+            this.percentage = Math.round(Number(this.percentage)).toString();
     }
 
 }
