@@ -2,5 +2,5 @@ export const environment = {
   production: true
 };
 
-export const API_URL = 'http://lb-enarmapi-599869600.us-east-1.elb.amazonaws.com/'
+export const API_URL = 'https://api.medbrain.click'
 
