@@ -2,4 +2,4 @@ export const environment = {
     production: false
 };
 
-export const API_URL = 'http://3.230.92.106:3000'
+export const API_URL = 'https://api.medbrain.click/'
