@@ -2,15 +2,15 @@ const config = {
     db: {
       /* don't expose password or any sensitive info, done only for demo */
       // aws rds
-      // host: "medbrain-dev.cxcex64atlct.us-east-1.rds.amazonaws.com",
-      // user: "admin",
-      // password: "7ZicQyUwv0aNGollri1H",
-      // database: "enarm_db",
-      // local
-      host: "127.0.0.1",
-      user: "root",
-      password: "enarm_db",
+      host: "medbrain-dev.cxcex64atlct.us-east-1.rds.amazonaws.com",
+      user: "admin",
+      password: "7ZicQyUwv0aNGollri1H",
       database: "enarm_db",
+      // local
+      // host: "127.0.0.1",
+      // user: "root",
+      // password: "Mex5todom!",
+      // database: "enarm_db",
     },
     listPerPage: 10,
   };
