@@ -4,6 +4,8 @@ const authService = require('../services/auth');
 const jwt = require('jsonwebtoken');
 var http = require('http2').constants;
 
+// const redisClient = require('../libs/redis');
+
 
 router.post('/login', async function (req, res, next) {
     try {
@@ -89,7 +91,38 @@ router.put('/change-account-status', async function (req, res, next) {
     }
 });
 
-router.put('/reset_password', async function (req, res, next) {
+router.post('/logout', async function (req, res, next) {
+    try {
+        const bearerHeader = req.headers['authorization'];
+        if (!bearerHeader) {
+            return res.status(401).json({ message: 'No autorizado' });
+        }
+
+        const bearer = bearerHeader.split(' ');
+        const bearerToken = bearer[1];
+        req.token = bearerToken;
+
+        const decoded = jwt.verify(req.token, 'secretkey');
+        console.log('decoded', decoded, decoded.data.data.data.email);
+        
+        const id = decoded.data.data.data.email;
+        
+        // Elimina la clave en Redis
+        const resultado = await redisClient.del(id);
+        
+        if (resultado) {
+            res.status(200).json({ message: 'Logout exitoso' });
+        } else {
+            res.status(404).json({ message: 'Sesión no encontrada' });
+        }
+    } catch (err) {
+        console.error(`Error en logout:`, err.message);
+        res.status(500).json({ message: 'Error interno del servidor' });
+        next(err);
+    }
+});
+
+router.put('/reset-password', async function (req, res, next) {
     try {
         const id = req.body.id;
         const resetPass = await authService.resetPassword(id)
