@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const exam = require('../services/exam');
 const { verifyToken } = require('../libs/headers');
+// const sessionValidator = require('../libs/session');
 
 router.post('/add-type', verifyToken, async function(req, res, next) {
   try {
@@ -29,7 +30,6 @@ router.post('/add-exam', verifyToken, async function(req, res, next) {
 
 router.put('/save-answer', verifyToken, async function (req, res, next) {
   try {
-
     const body = req.body;
     const response = await exam.saveAnswer(body)
     console.log('res', response);
